requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
backports.csv==1.0.6
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
cffi==1.11.5
chardet==3.0.4
cheroot==6.5.4
CherryPy==18.1.0
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
Django==2.1.5
django-autocomplete-light==3.3.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-multiselectfield==0.1.8
docker-py==1.10.6
docker-pycreds==0.4.0
feedparser==5.2.1
future==0.17.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jaraco.functools==2.0
lxml==4.3.0
more-itertools==5.0.0
mysqlclient==1.4.1
nltk==3.4
numpy==1.16.0
pandas==0.24.0
parsel==1.5.1
Pattern==3.6
pdfminer.six==20181108
pluggy==0.8.1
portend==2.3
psycopg2==2.7.7
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
pycryptodome==3.7.3
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pytest==4.1.1
pytest-runner==4.2
python-dateutil==2.7.5
python-docx==0.8.10
pytz==2018.9
PyYAML==3.13
queuelib==1.5.0
requests==2.21.0
retrying==1.3.3
scipy==1.2.0
scrapinghub==2.1.0
Scrapy==1.5.1
scrapy-splash==0.7.2
service-identity==18.1.0
shub==2.8.2
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.7.3
tempora==1.14
tqdm==4.30.0
urllib3==1.24.1
w3lib==1.20.0
websocket-client==0.54.0
wincertstore==0.2
zc.lockfile==1.4
zope.interface==4.6.0
